--- src/main/java/de/benevolo/member_management/membership/repositories/MembershipRepository.java
@@ -1,12 +1,0 @@
-package de.benevolo.member_management.membership.repositories;
-
-import de.benevolo.entities.association.Membership;
-import io.quarkus.hibernate.orm.panache.PanacheRepository;
-
-import javax.enterprise.context.ApplicationScoped;
-
-//Turn Bean ino a CDI Bean
-@ApplicationScoped
-public class MembershipRepository implements PanacheRepository<Membership> {
- // auto implemented
-}
