+++ src/main/java/de/benevolo/member_management/membership/services/MembershipService.java
@@ -1,13 +1,10 @@
 package de.benevolo.member_management.membership.services;
 
+import java.util.List;
+
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.transaction.Transactional;
-import javax.ws.rs.NotAcceptableException;
-
-import javax.json.Json;
-import javax.json.JsonObject;
-import javax.json.JsonObjectBuilder;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,7 +13,6 @@ import de.benevolo.entities.association.Association;
 import de.benevolo.entities.association.AssociationRole;
 import de.benevolo.entities.association.Membership;
 import de.benevolo.entities.user.PlatformUser;
-import de.benevolo.member_management.membership.repositories.MembershipRepository;
 
 import de.benevolo.member_management.membership.connectors.AssociationRestClient;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
@@ -27,58 +23,58 @@ public class MembershipService {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(MembershipService.class);
 
-    /* 
-    @Inject
-    MembershipRepository membershipRepository;
-    */
-
     @Inject
     @RestClient
     AssociationRestClient associationRestClient;
 
-
     @Transactional
-    public Boolean persistMembership(Association association, PlatformUser user){
+    public Boolean persistMembership(Association association, PlatformUser user) {
 
+        // Mocking a little bit
         Membership membership = new Membership(null, user.getId(), new AssociationRole());
         membership.setMembershipId(null);
         membership.getAssociationRole().setAssociationRoleId(null);
-        
-       
+
         try {
-            //membershipRepository.persist(membership);
+
             associationRestClient.addMembership(membership);
-            LOGGER.info("PERSISTET!" + membership.toString());
-           return true;
+            LOGGER.info("PERSISTET -->  " + membership.toString());
+            return true;
         } catch (Exception e) {
-           // logger.error("Could Not Persist Membership", e);
-           LOGGER.info("EXCEPTION" + e);
+
+            LOGGER.error("EXCEPTION -->  " + e);
             return false;
         }
 
-        //JSON Builder
+    }
 
-            
-/* 
-            JsonObjectBuilder roleBuilder = Json.createObjectBuilder();
-            roleBuilder.add("roleName", association.getAssociationRoles().getRoleName());
-            roleBuilder.add("roleDescription", association.getAssociationRoles().getRoleDescription());
-            roleBuilder.add("canTransaction", association.getAssociationRoles().getCanTransaction());
-            roleBuilder.add("canAccount", association.getAssociationRoles().getCanAccount());
-            roleBuilder.add("canEvent", association.getAssociationRoles().getCanEvent());
-            
-            JsonObjectBuilder jsonObjectBuilder = Json.createObjectBuilder();
-            jsonObjectBuilder.add("userId", user.getId());
-            jsonObjectBuilder.add("associationRole", roleBuilder);
-           
-            
-            JsonObject jsonObject = jsonObjectBuilder.build();
-            String json = jsonObject.toString();
-            
-            System.out.println(json);
+    public Boolean checkStatus(PlatformUser user, Association association) {
 
-*/
-        
+        Boolean result = false;
+
+        try {
+
+            List<Membership> members = associationRestClient.membersByAssociationId(association.getId());
+
+            LOGGER.info("Username: " + user.getName());
+            LOGGER.info("UserId: " + user.getId());
+            LOGGER.info("Generated List: " + members);
+
+            if (members.isEmpty() == true) {
+                LOGGER.info("Generated List EMPTY?: " + members.isEmpty());
+            } else {
+                for (Membership member : members) {
+                    LOGGER.info("MeberId: " + member.getUserId());
+                    if (member.getUserId() == user.getId()) {
+                        result = true;
+                    }
+                }
+            }
+
+        } catch (Exception e) {
+            LOGGER.error("Exception -> Error: Something wrong in fetching Data from Memberships");
+        }
+        return result;
     }
 
 }
