--- src/main/resources/update-database.bpmn2
@@ -1,92 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<bpmn2:definitions xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:bpmn2="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:bpsim="http://www.bpsim.org/schemas/1.0" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:drools="http://www.jboss.org/drools" id="_cV55sfJuEDu2hMwYxcpAGQ" xsi:schemaLocation="http://www.omg.org/spec/BPMN/20100524/MODEL BPMN20.xsd http://www.jboss.org/drools drools.xsd http://www.bpsim.org/schemas/1.0 bpsim.xsd http://www.omg.org/spec/DD/20100524/DC DC.xsd http://www.omg.org/spec/DD/20100524/DI DI.xsd " exporter="jBPM Process Modeler" exporterVersion="2.0" targetNamespace="http://www.omg.org/bpmn20">
-  <bpmn2:collaboration id="_E8BFBF1F-C301-440F-8F91-BA8A8EDED841" name="Default Collaboration">
-    <bpmn2:participant id="_AB75FBEA-1EB4-491C-8255-5B0A9E3FEF82" name="Pool Participant" processRef="UpdateDBafterCancelation"/>
-  </bpmn2:collaboration>
-  <bpmn2:process id="UpdateDBafterCancelation" drools:packageName="de.benevolo" drools:version="1.0" drools:adHoc="false" name="UpdateDBafterCancelation" isExecutable="true" processType="Public">
-    <bpmn2:laneSet>
-      <bpmn2:lane id="_94FF6136-4442-4C75-B6D3-162152BE11CB" name="UpdateDBAfterCancelation">
-        <bpmn2:extensionElements>
-          <drools:metaData name="elementname">
-            <drools:metaValue><![CDATA[UpdateDBAfterCancelation]]></drools:metaValue>
-          </drools:metaData>
-        </bpmn2:extensionElements>
-        <bpmn2:flowNodeRef>_96D7182C-611D-4C89-8DC3-09839A0C8C45</bpmn2:flowNodeRef>
-        <bpmn2:flowNodeRef>_19ADFE73-550F-49AB-9806-7F220B04945B</bpmn2:flowNodeRef>
-        <bpmn2:flowNodeRef>_393DBAC5-04EA-4585-B0B5-1CDA17AB03C0</bpmn2:flowNodeRef>
-      </bpmn2:lane>
-    </bpmn2:laneSet>
-    <bpmn2:sequenceFlow id="_D7BB3767-DE19-4F4B-9B5E-3D0BC8B1D57F" sourceRef="_19ADFE73-550F-49AB-9806-7F220B04945B" targetRef="_393DBAC5-04EA-4585-B0B5-1CDA17AB03C0"/>
-    <bpmn2:sequenceFlow id="_ED57924B-A3D5-4527-B6F7-EFBD67F0C01F" sourceRef="_96D7182C-611D-4C89-8DC3-09839A0C8C45" targetRef="_19ADFE73-550F-49AB-9806-7F220B04945B"/>
-    <bpmn2:endEvent id="_393DBAC5-04EA-4585-B0B5-1CDA17AB03C0" name="DB Update Finished">
-      <bpmn2:extensionElements>
-        <drools:metaData name="elementname">
-          <drools:metaValue><![CDATA[DB Update Finished]]></drools:metaValue>
-        </drools:metaData>
-      </bpmn2:extensionElements>
-      <bpmn2:incoming>_D7BB3767-DE19-4F4B-9B5E-3D0BC8B1D57F</bpmn2:incoming>
-    </bpmn2:endEvent>
-    <bpmn2:intermediateCatchEvent id="_19ADFE73-550F-49AB-9806-7F220B04945B" name="Wait / Implementation">
-      <bpmn2:extensionElements>
-        <drools:metaData name="elementname">
-          <drools:metaValue><![CDATA[Wait / Implementation]]></drools:metaValue>
-        </drools:metaData>
-      </bpmn2:extensionElements>
-      <bpmn2:incoming>_ED57924B-A3D5-4527-B6F7-EFBD67F0C01F</bpmn2:incoming>
-      <bpmn2:outgoing>_D7BB3767-DE19-4F4B-9B5E-3D0BC8B1D57F</bpmn2:outgoing>
-      <bpmn2:timerEventDefinition>
-        <bpmn2:timeDuration xsi:type="bpmn2:tFormalExpression">PT60S</bpmn2:timeDuration>
-      </bpmn2:timerEventDefinition>
-    </bpmn2:intermediateCatchEvent>
-    <bpmn2:startEvent id="_96D7182C-611D-4C89-8DC3-09839A0C8C45" name="Update DB">
-      <bpmn2:extensionElements>
-        <drools:metaData name="elementname">
-          <drools:metaValue><![CDATA[Update DB]]></drools:metaValue>
-        </drools:metaData>
-      </bpmn2:extensionElements>
-      <bpmn2:outgoing>_ED57924B-A3D5-4527-B6F7-EFBD67F0C01F</bpmn2:outgoing>
-    </bpmn2:startEvent>
-  </bpmn2:process>
-  <bpmndi:BPMNDiagram>
-    <bpmndi:BPMNPlane bpmnElement="UpdateDBafterCancelation">
-      <bpmndi:BPMNShape id="shape__96D7182C-611D-4C89-8DC3-09839A0C8C45" bpmnElement="_96D7182C-611D-4C89-8DC3-09839A0C8C45">
-        <dc:Bounds height="56" width="56" x="316" y="481"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="shape__19ADFE73-550F-49AB-9806-7F220B04945B" bpmnElement="_19ADFE73-550F-49AB-9806-7F220B04945B">
-        <dc:Bounds height="56" width="56" x="520" y="481"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="shape__393DBAC5-04EA-4585-B0B5-1CDA17AB03C0" bpmnElement="_393DBAC5-04EA-4585-B0B5-1CDA17AB03C0">
-        <dc:Bounds height="56" width="56" x="730" y="481"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNShape id="shape__94FF6136-4442-4C75-B6D3-162152BE11CB" bpmnElement="_94FF6136-4442-4C75-B6D3-162152BE11CB">
-        <dc:Bounds height="250" width="650" x="223" y="384"/>
-      </bpmndi:BPMNShape>
-      <bpmndi:BPMNEdge id="edge_shape__96D7182C-611D-4C89-8DC3-09839A0C8C45_to_shape__19ADFE73-550F-49AB-9806-7F220B04945B" bpmnElement="_ED57924B-A3D5-4527-B6F7-EFBD67F0C01F">
-        <di:waypoint x="344" y="509"/>
-        <di:waypoint x="548" y="509"/>
-      </bpmndi:BPMNEdge>
-      <bpmndi:BPMNEdge id="edge_shape__19ADFE73-550F-49AB-9806-7F220B04945B_to_shape__393DBAC5-04EA-4585-B0B5-1CDA17AB03C0" bpmnElement="_D7BB3767-DE19-4F4B-9B5E-3D0BC8B1D57F">
-        <di:waypoint x="548" y="509"/>
-        <di:waypoint x="758" y="509"/>
-      </bpmndi:BPMNEdge>
-    </bpmndi:BPMNPlane>
-  </bpmndi:BPMNDiagram>
-  <bpmn2:relationship type="BPSimData">
-    <bpmn2:extensionElements>
-      <bpsim:BPSimData>
-        <bpsim:Scenario id="default" name="Simulationscenario">
-          <bpsim:ScenarioParameters/>
-          <bpsim:ElementParameters elementRef="_96D7182C-611D-4C89-8DC3-09839A0C8C45">
-            <bpsim:TimeParameters>
-              <bpsim:ProcessingTime>
-                <bpsim:NormalDistribution mean="0" standardDeviation="0"/>
-              </bpsim:ProcessingTime>
-            </bpsim:TimeParameters>
-          </bpsim:ElementParameters>
-        </bpsim:Scenario>
-      </bpsim:BPSimData>
-    </bpmn2:extensionElements>
-    <bpmn2:source>_cV55sfJuEDu2hMwYxcpAGQ</bpmn2:source>
-    <bpmn2:target>_cV55sfJuEDu2hMwYxcpAGQ</bpmn2:target>
-  </bpmn2:relationship>
-</bpmn2:definitions>
